wallet refactor; deeplink changed; utils refactor
