+++ lib/account_dialog/widget/pages/0_main.dart
@@ -6,7 +6,7 @@ import '../../../blockchain/interface.dart';
 import '../../../blockchain/task_services.dart';
 import '../../../config/theme.dart';
 import '../../../widgets/badge-small-colored.dart';
-import '../../../widgets/utils/my_tools.dart';
+import '../../../config/utils/my_tools.dart';
 
 import 'package:badges/badges.dart' as Badges;
