--- lib/config/flutter_flow_util.dart
@@ -1,200 +1,0 @@
-import 'dart:io';
-
-import 'package:flutter/foundation.dart' show kIsWeb;
-import 'package:flutter/material.dart';
-import 'package:intl/intl.dart';
-import 'package:json_path/json_path.dart';
-import 'package:timeago/timeago.dart' as timeago;
-import 'package:url_launcher/url_launcher.dart';
-
-import '../main.dart';
-
-export 'dart:math' show min, max;
-export 'dart:convert' show jsonEncode, jsonDecode;
-export 'package:intl/intl.dart';
-export 'package:page_transition/page_transition.dart';
-
-T valueOrDefault<T>(T? value, T defaultValue) =>
-    (value is String && value.isEmpty) || value == null ? defaultValue : value;
-
-String dateTimeFormat(String format, DateTime? dateTime) {
-  if (dateTime == null) {
-    return '';
-  }
-  if (format == 'relative') {
-    return timeago.format(dateTime);
-  }
-  return DateFormat(format).format(dateTime);
-}
-
-Future launchURL(String url) async {
-  var uri = Uri.parse(url).toString();
-  try {
-    await launch(uri);
-  } catch (e) {
-    throw 'Could not launch $uri: $e';
-  }
-}
-
-enum FormatType {
-  decimal,
-  percent,
-  scientific,
-  compact,
-  compactLong,
-  custom,
-}
-
-enum DecimalType {
-  automatic,
-  periodDecimal,
-  commaDecimal,
-}
-
-String formatNumber(
-  num? value, {
-  required FormatType formatType,
-  DecimalType? decimalType,
-  String? currency,
-  bool toLowerCase = false,
-  String? format,
-  String? locale,
-}) {
-  var formattedValue = '';
-  switch (formatType) {
-    case FormatType.decimal:
-      switch (decimalType!) {
-        case DecimalType.automatic:
-          formattedValue = NumberFormat.decimalPattern().format(value);
-          break;
-        case DecimalType.periodDecimal:
-          formattedValue = NumberFormat.decimalPattern('en_US').format(value);
-          break;
-        case DecimalType.commaDecimal:
-          formattedValue = NumberFormat.decimalPattern('es_PA').format(value);
-          break;
-      }
-      break;
-    case FormatType.percent:
-      formattedValue = NumberFormat.percentPattern().format(value);
-      break;
-    case FormatType.scientific:
-      formattedValue = NumberFormat.scientificPattern().format(value);
-      if (toLowerCase) {
-        formattedValue = formattedValue.toLowerCase();
-      }
-      break;
-    case FormatType.compact:
-      formattedValue = NumberFormat.compact().format(value);
-      break;
-    case FormatType.compactLong:
-      formattedValue = NumberFormat.compactLong().format(value);
-      break;
-    case FormatType.custom:
-      final hasLocale = locale != null && locale.isNotEmpty;
-      formattedValue =
-          NumberFormat(format, hasLocale ? locale : null).format(value);
-  }
-
-  if (formattedValue.isEmpty) {
-    return value.toString();
-  }
-
-  if (currency != null) {
-    final currencySymbol = currency.isNotEmpty
-        ? currency
-        : NumberFormat.simpleCurrency().format(0.0).substring(0, 1);
-    formattedValue = '$currencySymbol$formattedValue';
-  }
-
-  return formattedValue;
-}
-
-DateTime get getCurrentTimestamp => DateTime.now();
-
-extension DateTimeComparisonOperators on DateTime {
-  bool operator <(DateTime other) => isBefore(other);
-  bool operator >(DateTime other) => isAfter(other);
-  bool operator <=(DateTime other) => this < other || isAtSameMomentAs(other);
-  bool operator >=(DateTime other) => this > other || isAtSameMomentAs(other);
-}
-
-dynamic getJsonField(dynamic response, String jsonPath) {
-  final field = JsonPath(jsonPath).read(response);
-  return field.isNotEmpty
-      ? field.length > 1
-          ? field.map((f) => f.value).toList()
-          : field.first.value
-      : null;
-}
-
-bool get isAndroid => !kIsWeb && Platform.isAndroid;
-bool get isiOS => !kIsWeb && Platform.isIOS;
-bool get isWeb => kIsWeb;
-bool responsiveVisibility({
-  required BuildContext context,
-  bool phone = true,
-  bool tablet = true,
-  bool tabletLandscape = true,
-  bool desktop = true,
-}) {
-  final width = MediaQuery.of(context).size.width;
-  if (width < 479) {
-    return phone;
-  } else if (width < 767) {
-    return tablet;
-  } else if (width < 991) {
-    return tabletLandscape;
-  } else {
-    return desktop;
-  }
-}
-
-const kTextValidatorUsernameRegex = r'^[a-zA-Z][a-zA-Z0-9_-]{2,16}$';
-const kTextValidatorEmailRegex =
-    r"^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,253}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,253}[a-zA-Z0-9])?)*$";
-const kTextValidatorWebsiteRegex =
-    r'(https?:\/\/)?(www\.)[-a-zA-Z0-9@:%._\+~#=]{2,256}\.[a-z]{2,4}\b([-a-zA-Z0-9@:%_\+.~#?&//=]*)|(https?:\/\/)?(www\.)?(?!ww)[-a-zA-Z0-9@:%._\+~#=]{2,256}\.[a-z]{2,4}\b([-a-zA-Z0-9@:%_\+.~#?&//=]*)';
-
-void setAppLanguage(BuildContext context, String language) =>
-    MyApp.of(context).setLocale(Locale(language, ''));
-
-void setDarkModeSetting(BuildContext context, ThemeMode themeMode) =>
-    MyApp.of(context).setThemeMode(themeMode);
-
-void showSnackbar(
-  BuildContext context,
-  String message, {
-  bool loading = false,
-  int duration = 4,
-}) {
-  ScaffoldMessenger.of(context).hideCurrentSnackBar();
-  ScaffoldMessenger.of(context).showSnackBar(
-    SnackBar(
-      content: Row(
-        children: [
-          if (loading)
-            Padding(
-              padding: EdgeInsetsDirectional.only(end: 10.0),
-              child: Container(
-                height: 20,
-                width: 20,
-                child: const CircularProgressIndicator(
-                  color: Colors.white,
-                ),
-              ),
-            ),
-          Text(message),
-        ],
-      ),
-      duration: Duration(seconds: duration),
-    ),
-  );
-}
-
-extension FFStringExt on String {
-  String maybeHandleOverflow({int? maxChars, String replacement = ''}) =>
-      maxChars != null && length > maxChars
-          ? replaceRange(maxChars, null, replacement)
-          : this;
-}
