+++ lib/config/utils/util.dart
@@ -1,0 +1,49 @@
+import 'dart:io';
+
+import 'package:flutter/foundation.dart' show kIsWeb;
+import 'package:flutter/material.dart';
+import 'package:intl/intl.dart';
+import 'package:json_path/json_path.dart';
+import 'package:timeago/timeago.dart' as timeago;
+import 'package:url_launcher/url_launcher.dart';
+
+import '../../main.dart';
+
+export 'dart:math' show min, max;
+export 'dart:convert' show jsonEncode, jsonDecode;
+export 'package:intl/intl.dart';
+export 'package:page_transition/page_transition.dart';
+
+String dateTimeFormat(String format, DateTime? dateTime) {
+  if (dateTime == null) {
+    return '';
+  }
+  if (format == 'relative') {
+    return timeago.format(dateTime);
+  }
+  return DateFormat(format).format(dateTime);
+}
+
+Future launchURL(String url) async {
+  var uri = Uri.parse(url).toString();
+  try {
+    await launch(uri);
+  } catch (e) {
+    throw 'Could not launch $uri: $e';
+  }
+}
+
+
+extension DateTimeComparisonOperators on DateTime {
+  bool operator <(DateTime other) => isBefore(other);
+  bool operator >(DateTime other) => isAfter(other);
+  bool operator <=(DateTime other) => this < other || isAtSameMomentAs(other);
+  bool operator >=(DateTime other) => this > other || isAtSameMomentAs(other);
+}
+
+
+const kTextValidatorUsernameRegex = r'^[a-zA-Z][a-zA-Z0-9_-]{2,16}$';
+const kTextValidatorEmailRegex =
+    r"^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,253}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,253}[a-zA-Z0-9])?)*$";
+const kTextValidatorWebsiteRegex =
+    r'(https?:\/\/)?(www\.)[-a-zA-Z0-9@:%._\+~#=]{2,256}\.[a-z]{2,4}\b([-a-zA-Z0-9@:%_\+.~#?&//=]*)|(https?:\/\/)?(www\.)?(?!ww)[-a-zA-Z0-9@:%._\+~#=]{2,256}\.[a-z]{2,4}\b([-a-zA-Z0-9@:%_\+.~#?&//=]*)';
