+++ lib/pages/auditor_page.dart
@@ -16,9 +16,6 @@ import '../task_dialog/main.dart';
 import '../widgets/loading.dart';
 import '../widgets/tags/main.dart';
 import '../widgets/tags/search_services.dart';
-import '../widgets/tags/tag_open_container.dart';
-import '../task_item/task_item.dart';
-import '../config/flutter_flow_animations.dart';
 import '../config/theme.dart';
 import 'package:flutter/material.dart';
 import 'package:beamer/beamer.dart';
