+++ lib/task_dialog/pages/2_description.dart
@@ -7,7 +7,7 @@ import 'package:webthree/credentials.dart';
 import '../../blockchain/interface.dart';
 import '../../blockchain/classes.dart';
 import '../../blockchain/task_services.dart';
-import '../../config/flutter_flow_util.dart';
+import '../../config/utils/util.dart';
 import '../../config/theme.dart';
 import '../widget/dialog_button_widget.dart';
 import '../widget/request_audit_alert.dart';
