+++ lib/task_item/task_item.dart
@@ -7,14 +7,11 @@ import 'package:shimmer/shimmer.dart';
 
 import '../blockchain/classes.dart';
 
-import '../blockchain/interface.dart';
 import '../blockchain/task_services.dart';
-import '../config/flutter_flow_util.dart';
+import '../config/utils/util.dart';
 import '../config/theme.dart';
-import '../config/flutter_flow_util.dart';
 import '../wallet/model_view/wallet_model.dart';
 import '../wallet/services/wallet_service.dart';
-import '../widgets/tags/tags_old.dart';
 import '../widgets/tags/wrapped_chip.dart';
 import '../widgets/delete_item_alert.dart';
