+++ lib/wallet/widgets/pages/2_wc/main.dart
@@ -1,14 +1,10 @@
 
-import 'package:dodao/wallet/widgets/pages/2_wc/tab_mobile_desktop.dart';
-import 'package:dodao/wallet/widgets/pages/2_wc/tab_qr_code.dart';
-import 'package:dodao/wallet/widgets/pages/2_wc/wc_action_button.dart';
+import 'package:dodao/wallet/widgets/pages/2_wc/screen_states.dart';
+import 'package:dodao/wallet/widgets/pages/2_wc/action_button.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-
-import '../../../../config/theme.dart';
 import '../../../model_view/wc_model.dart';
-import 'connect_disconnect_state.dart';
 
 class WalletConnectMainScreen extends StatelessWidget {
   const WalletConnectMainScreen({
@@ -25,7 +21,7 @@ class WalletConnectMainScreen extends StatelessWidget {
     return Column(
       crossAxisAlignment: CrossAxisAlignment.center,
       children: [
-        WcQrCodeTab(screenHeightSizeNoKeyboard: screenHeightSizeNoKeyboard),
+        WCStates(screenHeightSizeNoKeyboard: screenHeightSizeNoKeyboard),
         const Spacer(),
         // const ConnectDisconnectState(),
         const WCActionButton(),
@@ -80,7 +76,7 @@ class WalletConnectMainScreen extends StatelessWidget {
     //             Column(
     //               crossAxisAlignment: CrossAxisAlignment.center,
     //               children: [
-    //                 WcQrCodeTab(screenHeightSizeNoKeyboard: screenHeightSizeNoKeyboard),
+    //                 WCStates(screenHeightSizeNoKeyboard: screenHeightSizeNoKeyboard),
     //                 const Spacer(),
     //                 const ConnectDisconnectState(),
     //                 const WCActionButton(),
